--- lib/src/main/java/debug_artist/menu/drawer/item/input/InputItemListener.kt
@@ -1,5 +1,0 @@
-package debug_artist.menu.drawer.item.input
-
-interface InputItemListener {
-  fun onTextInputEnter(itemId: Int, inputText: String)
-}
