--- lib/src/main/java/debug_artist/menu/drawer/item/spinner/SpinnerItemListener.kt
@@ -1,5 +1,0 @@
-package debug_artist.menu.drawer.item.spinner
-
-interface SpinnerItemListener {
-  fun onSpinnerItemClick(item: SpinnerDrawerItem, itemId: Int, title: CharSequence)
-}
