+++ lib/src/main/java/debugartist/menu/drawer/item/input/InputItemListener.kt
@@ -1,0 +1,5 @@
+package debugartist.menu.drawer.item.input
+
+interface InputItemListener {
+  fun onTextInputEnter(itemId: Int, inputText: String)
+}
