+++ lib/src/main/java/debugartist/menu/drawer/item/spinner/SpinnerItemListener.kt
@@ -1,0 +1,5 @@
+package debugartist.menu.drawer.item.spinner
+
+interface SpinnerItemListener {
+  fun onSpinnerItemClick(item: SpinnerDrawerItem, itemId: Int, title: CharSequence)
+}
