+++ sample/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="debug_artist.sample">
+  package="debugartist.sample">
 
   <application
     android:allowBackup="true"
@@ -9,7 +9,7 @@
     android:theme="@style/AppTheme">
 
     <activity
-      android:name=".LauncherActivity"
+      android:name="debugartist.sample.LauncherActivity"
       android:label="@string/sample">
       <intent-filter>
         <action android:name="android.intent.action.MAIN" />
@@ -20,7 +20,7 @@
     </activity>
 
     <activity
-      android:name=".base.BasicActivity"
+      android:name="debugartist.sample.base.BasicActivity"
       android:label="Basic Sample" />
 
   </application>
